+++ capskip/__init__.py
@@ -26,4 +26,4 @@ __all__ = [
     'TimeoutException',
 ]
 
-__version__ = '1.0.1'
+__version__ = '1.0.2'